#9 delete except - appropriate implementation

+++ src/repositories/device-auth-sessions-repository.ts
@@ -26,8 +26,8 @@ export const deviceAuthSessionsRepository = {
     async deleteAll() {
         await deviceAuthSessionsCollection.deleteMany({})
     },
-    async deleteExcept(deviceId: string) {
-        await deviceAuthSessionsCollection.deleteMany({deviceId: {$ne: deviceId}})
+    async deleteExcept(userId: ObjectId, deviceId: string) {
+        await deviceAuthSessionsCollection.deleteMany({$and: [{userId: userId}, {deviceId: {$ne: deviceId}}]})
     },
     async deleteByDeviceId(deviceId: string) {
         await deviceAuthSessionsCollection.deleteOne({deviceId: deviceId})
